Started implementing class for unit cube
