--- COMP371-Assignment1.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29709.97
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "COMP371-Assignment1", "Labs.vcxproj", "{277EAE68-9EED-4EFF-BAD3-2A70787A3E8D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Win32 = Debug|Win32
-		Release|Win32 = Release|Win32
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{277EAE68-9EED-4EFF-BAD3-2A70787A3E8D}.Debug|Win32.ActiveCfg = Debug|Win32
-		{277EAE68-9EED-4EFF-BAD3-2A70787A3E8D}.Debug|Win32.Build.0 = Debug|Win32
-		{277EAE68-9EED-4EFF-BAD3-2A70787A3E8D}.Release|Win32.ActiveCfg = Release|Win32
-		{277EAE68-9EED-4EFF-BAD3-2A70787A3E8D}.Release|Win32.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {EF2BDAF3-767B-4B12-ACF9-B67DB3AB1A30}
-	EndGlobalSection
-EndGlobal
